Convert locations to design-coordinates to fix instance adding
Instance positions were specified in user-coordinates but were added to
the designspace as design-coordinates without conversion, which led some
to become out-of-bounds and be dropped by the build process.

This commit adds an explicit conversion to design-coordinates - ensuring
that instances' stay within bounds - and documents which coordinate
space the locations are in.