Keep going if we cannot read some UFOs and glifs, to handle WIP commits
Previously, if a config or designspace file could not be read, it would
be skipped, but a UFO or glif file that could not be parsed would cause
crashes.

This commit extends the same skipping behaviour to UFOs and glif files,
as the burndown generator should expect to encounter git histories
containing WIP commits, and should still produce a graph for them.

The current implicit behaviour is to show these glyphs as having no
status, which would be visible as a dip on the graph, but this could be
adjusted if unhelpful.